requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0

# Twitter API
twarc>=2.14.0
TwitterAPI>=3.0.0

# NLP and Machine Learning
transformers>=4.21.0
torch>=1.12.0
nltk>=3.7
scikit-learn>=1.1.0

# Translation
googletrans==4.0.0rc1

# Image Processing
opencv-python>=4.6.0
Pillow>=9.2.0
pytesseract>=0.3.10
deepface>=0.0.79

# Data Processing
requests>=2.28.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0

# Utilities
python-dotenv>=0.20.0

